fix: resolve 12 console errors + optimize homepage
Console errors fixed:
- manifest.json: use absolute path /webapp/manifest.json
- /admin/track 401: skip tracking when user not logged in
- /auth/user 401: add silent:true to suppress error toast
- Media thumbnails 401: make thumb/cover paths publicly accessible

Homepage optimization:
- Redesigned hero section with feature highlights (HLS, resumable upload, secure sharing, self-hosted)
- Removed forced AuthDialog for guests, replaced with gentle CTA
- Added responsive 2-column feature grid with hover effects
- i18n strings for both zh-CN and en-US

Backend:
- cargo fmt (formatting cleanup across codebase)
- cargo clippy --fix (auto-fix lint issues)

+++ backend/src/models/admin.rs
@@ -65,7 +65,7 @@ pub struct TranscodeRequest {
 pub struct TranscodeResponse {
     pub success: bool,
     pub message: String,
-    pub job_id: Option<i32>,
+    pub job_id: Option<i64>,
 }
 
 #[derive(Serialize)]
